fix:validation and change migration int to bigint

+++ database/migrations/2023_11_18_054102_create_transactions.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->id();
-            $table->integer("nominal");
+            $table->bigInteger("nominal");
             $table->date("tgl");
             $table->string("keterangan");
             $table->enum('status', ['pemasukan', 'pengeluaran']);
